Rework page structure
