--- edition-archive-frontend/pages/gallia_pontifica_online/[[path]].vue
@@ -1,36 +1,0 @@
-<template>
-  <GalliaPontificaOnlineLayout>
-    <template #content>
-        <div class="row">
-          <div class="col">
-            <ContentRenderer v-if="data" :value="data"/>
-          </div>
-        </div>
-    </template>
-  </GalliaPontificaOnlineLayout>
-</template>
-
-<script>
-  export default {}
-</script>
-
-<script setup>
-  const {path} = useRoute()
-  import {createError} from 'h3'
-
-  const {data, error} = await useAsyncData(`content-${path}`, () => {
-    const p1 = path.lastIndexOf("/") === path.length - 1 ? path + "beschreibung" : path;
-    return queryContent().where({_path: p1}).findOne()
-  })
-  if (error.value) {
-    throwError(
-        createError({
-          statusCode: 404,
-          statusMessage: 'Not Found',
-        })
-    );
-  }
-</script>
-
-<style scoped>
-</style>
