+++ edition-archive-frontend/pages/gallia_pontifica_online/projekt/[[projekt]]/regest/personen.vue
@@ -1,17 +1,20 @@
 <template>
   <GalliaPontificaOnlineLayout>
+
     <template #content>
-      <GalliaPontificaOnlineReimser>
-        <template #regesten>
           <h2>{{ $t("person_index") }}</h2>
           <ul class="list-group list-group-flush mt-5" v-if="data">
             <li v-for="doc in data.docs" class="list-group-item">
               <nuxt-link :href="`/gallia_pontifica_online/regest/suche/person/?personObj=${doc.id}`">{{ doc.displayName.join(", ") }}</nuxt-link>
             </li>
           </ul>
-        </template>
-      </GalliaPontificaOnlineReimser>
     </template>
+
+    <template #menu>
+      <GalliaPontificaOnlineReimserMenu v-if="$route.params.projekt == 'reimser'" />
+      <GalliaPontificaOnlineTherouanneMenu v-else />
+    </template>
+
   </GalliaPontificaOnlineLayout>
 </template>
